[ubsan] Fix various ClusterFuzz-found issues
Fixing a few float and int overflows.
Drive-by fix: with --experimental-wasm-bigint, Number values
may not be used to initialize i64-typed globals. The existing
code for doing that relied on UB; since it's a spec violation
the fix is to throw instead.

No regression test for 933103 because it will OOM anyway.
No regression test for 932896 because it would be extremely slow.

Bug: chromium:927894,chromium:927996,chromium:930086,chromium:932679,chromium:932896,chromium:933103,chromium:933134
Change-Id: Iae1c1ff1038af4512a52d3e56b8c4b75f2233314