Update patch set 1
Patch Set 1:

Thank you for your detailed reply. I agree that we should not get any dependencies to other parts of V8.

Is it guaranteed that the exception is a nullptr dereference if TLS is not initialized yet? Would it be possible to check for a nullptr exception before we do the thread-in-wasm check? The downside would be that if the nullptr check causes an exception, we cause another recursive crash.

Patch-set: 1
Reviewer: Gerrit User 1141154 <1141154@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>