Update patch set 18
Patch Set 18:

I don't understand yet where your verification problem is coming from. It seems to me that the graph you build initially should be valid (as it is the same graph you build for 64-bit systems), and also the graph after the lowering should be valid. Introducing a CallUnverified should not be necessary.

In the end what you have is originally one node (the call to the i64->bigint conversion) which has one input and one output. In the lowering phase you replace it with a new node which takes two nodes as inputs and produces one output. The same is happening for the I32SConvertI64 instruction, so in principle it is possible. It's also possible to change call nodes.

I admit I don't have a clear idea on how we can detect i64->bigint call nodes in the lowering phase, but I can't believe that that's so hard to do. I can ask some people on the team if that is really the problem.

Patch-set: 18
Reviewer: Gerrit User 1141154 <1141154@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>