Update patch set 18
Patch Set 18:

> Patch Set 18:
> 
> I don't understand yet where your verification problem is coming from. It seems to me that the graph you build initially should be valid (as it is the same graph you build for 64-bit systems), and also the graph after the lowering should be valid. Introducing a CallUnverified should not be necessary.
> 
> In the end what you have is originally one node (the call to the i64->bigint conversion) which has one input and one output. In the lowering phase you replace it with a new node which takes two nodes as inputs and produces one output. The same is happening for the I32SConvertI64 instruction, so in principle it is possible. It's also possible to change call nodes.
> 
> I admit I don't have a clear idea on how we can detect i64->bigint call nodes in the lowering phase, but I can't believe that that's so hard to do. I can ask some people on the team if that is really the problem.

The verification fails in different ways, I will try my best to explain it clearly because it's very confusing. I'm also happy to do a video chat.

For example, the builtin that does the i64->bigint conversion; in 64-bit its signature is one input and one ouput, and in 32-bit it's two inputs (pair of i32) and one ouput. Also, each builtin has its corresponding interface-descriptor. The calls to Parameter() in the builtin are staticly typed in a sense; they refer to parameters in the interface-descriptor. Given that and in my case, the builtin designed for 32-bit and the one designed for 64-bit are not interchangeable easily (because they refer to other parameters or arity of input), that's why I introduced the compilation flags and used only one interface-descriptor for both architectures.

Another issue can occur during the machine graph construction. The call to the i64->bigint builtin looks like that in 64-bit: `Call(target, Int64T, ...)` and for 32-bit: `Call(target, Int32T, Int32T, ...)`. In this CL, since the calling descriptor is "staticly" either 32-bit or 64-bit, the verification (that kicks in when you add a new node IRC) succeeded. Previsouly, it would have catch the issue even before reaching the lowering pass.

On a side note, as far as I can tell a compiled v8 in 32-bit isn't going to generate 64-bit code at runtime (apart from the simulator which doesn't seem to fail).

Patch-set: 18
Reviewer: Gerrit User 1229060 <1229060@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0