Fix Wasm trap handler recursion on exceptions raised early
Note: Putting this up for discussion, not a request to submit.

__declspec(thread) is problematic when used in a vectored exception
handler. The issue is that exceptions can get raised REALLY early in
the thread's life, even before the thread local storage for
"__declspec(thread)" variables is allocated. Should this occur, and the
handler tries to access a thread local variable, it will attempt to
access an offset from a null pointer, and raise another exception. The
new exception will cause this handler to get called again, and an infinite loop ensues.

In crash reports we were seeing recursion on the stack with the below:

 - msedge_child!v8::internal::trap_handler::HandleWasmTrap+0x27
 - ntdll!RtlpCallVectoredHandlers+0x106
 - ntdll!RtlCallVectoredExceptionHandlers+0xe
 - ntdll!RtlDispatchException+0x69
 - ntdll!KiUserExceptionDispatch+0x2e
 - msedge_child!v8::internal::trap_handler::HandleWasmTrap+0x27

There is not currently a simple and supported way to check if the
"implicit" thread-local variables, (i.e. those declared with
"__declspec(thread)"), have been initialized for the thread.

If thread local storage must be used in a vectored exception handler
(as here), then using the Win32 Tls* APIs is the best option. The
TlsGetValue API returns 0 on error, or if the slots haven't been
allocated yet (as they are 0 initialized). In my looking through the
code I could see this is already being done to store the current
isolate on enter/exit. Check this first seems like a nice clean and
simple approach before doing more advanced checks.

I'm new to this code though, so my approach may be entirely naive.

See the series of posts listed at http://www.nynaeve.net/?p=190 for
more detail than you ever wanted to know about thread locale storage
internals on Windows.

Change-Id: Ifc6223b74999a55bfd0ed2d6ebf054bbffd7e809