[wasm] Fix missing initialization
The array allocated in {SetInterpreterRedirection} was accidentally
left uninitialized, which could result in later calls to
{has_interpreter_redirection} to return false. In that case, we would
not patch the jump table when publishing the code, and would then crash
during execution.

This was uncovered by an unrelated change. It's covered by existing
tests though, we were just lucky so far that the relevant bits were 0
for some reason.

R=mstarzinger@chromium.org

Change-Id: I712200cb0c9f2a9c828f6a9a502e6307dc438bd3