[heap] Don't access flags from concurrent marking visitor
It is not safe to access flags from concurrent marking visitor. We access
FLAG_flush_bytecode and FLAG_stress_flush_bytecode when visiting
SharedFunctionInfo and JSFunction to decide if we need to collect bytecode.
This cl adds a bytecode_flushing_mode which will be initialized when creating
the visitor. This way we can avoid accessing flags.

Bug:v8:9045
Change-Id: I84bf09ec2dd1543abad54bd87f8bf953830b89e7