Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (4 comments)
> 
> Thanks!
> 
> First round of comments:
>   - Please verify the 'script' field in StackFrameInfoVerify in objects-debug.cc
>   - You will also need a +1 from an owner for src/heap/*
> 
> The overlap between StackFrameInfo and ScriptOrigin is a bit unfortunate. Alternatively we could incorporate all the necessary fields needed from ScriptOrigin into StackFrameInfo.

Thanks so much for the review, I've updated these.

Yes the overlap is unfortunate. I did initially look into seeing how the fields could be shared to construct a ScriptOrigin without a Script reference, but that would effectively end up with storing a lot more fields on StackFrameInfo, and I wasn't sure that there weren't subtle semantic differences with the shared ones as well.

I must admit to me, architecturally, converging on a single ScriptOrigin representation through the Script reference (and removing the redundant StackFrameInfo fields in the binary) feels simpler, and that that direction can be refactored more easily over time, instead of having different representations, but I don't want to over-generalize either without fully understanding the extend of the consumer side of these APIs.

Happy to discuss this further as well.

Patch-set: 3