Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> ping... any further concerns or questions on this? Thanks.

Hi, sorry for letting this CL languish.

We had a discussion in my team about whether this added complexity of the thread-local flag is worth the gain. As I recall the discussions from 3+ years ago, the idea was for this thread-local to increase confidence that the trap mechanism could not be hijacked. We had some concerns whether this mechanism did indeed accomplish this (e.g. assuming an attacker has a write, they could as well easily flip this flag).

I think we'd be OK landing this CL in the interim.

Is there a solution for Linux?

Patch-set: 5
CC: Gerrit User 1119499 <1119499@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>