DeoptimizationEntries code stubs could be builtins
Currently DeoptimizationEntries are generated as code stubs on a per-isolate
basis when the optimizer first optimizes some code. This CL transforms them
into embedded builtins.
This also fixes issue 8768: "Ensure we don't serialize generated
deopt_entry_code".

In the DeoptimizationEntries the root register was used to pass the bailout id.
This would not work when they become embedded builtins, so the bailout id is now
stored by the caller in the isolate data.

Bug: v8:8661
Change-Id: I572a18f66f61f7792e11a85180fff77766d95020