Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Just to clarify, what was changed from the original CL? The diff to PS3 lgtm.

It was a bad mistake, after all the refactoring I ended up with code in src\unwinding-info-win64.cc, CanRegisterUnwindInfoForNonABICompliantCodeRange() that returned true even when FLAG_win64_unwinding_info is false. 
Because of this, when an Isolate got initialized, RegisterNonABICompliantCodeRange() was called and even if it was not actually registering unwind data it still marking the first page of the code range as not writable. 
Then, when Chromium was being notified of the new isolate, it also tried to register its own unwind info for exception handling, and was crashing in components/crash/content/app/crashpad_win.cc, trying to write into that first page that it also reserves for unwind data.
The fix prevents RegisterNonABICompliantCodeRange() from being called when FLAG_win64_unwinding_info is not set.
Sorry...

Patch-set: 3
Reviewer: Gerrit User 1318771 <1318771@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0