[interpreter] Omit bounds check for interrupt budget loads in bytecode handlers
With lazy feedback allocation we load the interrupt budget from the
ClosureFeedbackCellArray instead of the bytecode array. The factory method
that constructs the ClosureFeedbackCellArray ensures we have a field for
interrupt budget. So, it is safe to omit bounds check here. Including the
bounds check increases the size of all jump bytecode handlers by around
~120 bytes. This translates to ~9-10KB of the native code size on Android.

Bug:chromium:948835
Change-Id: I8e10b3f630097728ae9e520bfb0d85dfc0f806ea