Update patch set 2
Patch Set 2:

Leszek, 

The bounds check itself does not need 120 bytes, but the generated code tries
to restore the state when the bounds check fails and we called abort. This is dead code and could be eliminated. I will file a different bug for that. Even without that I think it is Ok to omit the bounds check here. If the ClosureFeedbackCellArray wasn't constructed properly then there are so many other ways it can fail. PTAL.

Thanks,
Mythri

Patch-set: 2
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false