Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 2: Code-Review+1
> > > 
> > > I agree. I could imagine in the future this becoming an int field rather than a Smi field, in which case a bounds check would make even less sense.
> > 
> > I do not understand how anything can become an int field in FixedArray. Could you explain?
> 
> Make it not a FixedArray :) more specifically, do the same sort of thing as FeedbackVector already does, of having a fixed-size header with flexible pointers after it, a custom body descriptor, etc.

Yes, I considered doing that but just went with FixedArray. I am looking at Angular in speedometer2 and see two different jump instructions reasonably hot. So may be it is worth changing it so we could save a couple of instructions (both in size and runtime). I am not sure it moves the needle much, but doesn't add much complexity either.

Patch-set: 3