[api] allow any js value in host_defined_options
In Node.js we are having issues tracking values associated with
modules because we can only store data by primitive values.

This change allows V8 to handle the GC of data associated with
a script instead of forcing Node.js to perform its own tracking
which results in an unavoidable memory leak.

This was not a problem for Chromium because it does not expose
APIs like Node.js's VM which allow short lived scripts and modules.

Change-Id: Idccfc8144b2ef953adb9dc67966ac7c1b8b1fde7