Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > My memory of why this was a PrimitiveArray is vague (https://crbug.com/v8/6683 has some discussion but not a full explanation). But this is a major API change that I'm not sure we want (even if we do want to do something like this, we'll likely need a migration path from the old API to a new one).
> > > > 
> > > > Sathya & Georg, do you have better memory than I about HostDefinedOptions?
> > > 
> > > https://docs.google.com/document/u/1/d/1-8KW9vyuuXSLI6Mxo2oNqDb-Q-vlph0JDKC88k-rPwg/pub has more background. See the discussion under Version 3.
> > > 
> > > The reason we went with this PrimitiveArray is because Blink needed more than one value to be associated with a Script.
> > 
> > How important was it that these only be primitives, though? From the doc it seems like this was mostly matching Blink's requirements closely with as little overhead as possible.
> 
> We limited to primitives so that embedders wouldn't accidentally leak more memory by stashing objects here.

So the problem is that in the resolver APIs, we usually want some JS objects to retain the context of a module import parent. Eg the VM API does this in Node.js by providing the parent vm.Module instance. We're also thinking of doing similar in the new Loader API in providing eg a context object containing various data relating to the parent resolution.

Because of dynamic import (which was also one of the reasons for this host_data I believe), we can't know when to dispose this data.

Weak maps aren't sufficient because they do not permit symbols (https://github.com/tc39/ecma262/issues/1194#issuecomment-480075837).

So it seems what was designed to avoid a memory leak actually created an architectural one by not allowing the GC to apply to these cases.

Patch-set: 1