Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> This CL (https://chromium-review.googlesource.com/c/v8/v8/+/1416071) clears the microtask queue upon detaching of native context. Just looked into this together with Ulan. Is this a performance optimization? Or do we need this for correctness? It looks like this has tricky intersections with WeakRefs.

The cancellation is for the spec conformance, and the removal of the reference is for the lifetime management of MicrotaskQueue.
As MicrotaskQueue is a C++ object referred by Context, which is a GCed object, we have to clear the reference explicitly at some point to ensure that Context does not touch a stale object.

Patch-set: 4
Reviewer: Gerrit User 1211094 <1211094@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>