Update patch set 3
Patch Set 3:

What is the motivation for changing the stubs into builtins? Is it intended to be a code size improvement, as the builtins can be shared across isolates?

As it is, I am worried that this change can actually result in a code size increase for some cases (at least for some of the architectures).

For arm64, for example, the current stubs are 1092 bytes each and we have 3 of them, so 3276 bytes. Each call to the deoptimizer currently needs two instructions, so 8 bytes. The new code to call the deoptimizer will generate 10 instructions instead, so for each deopt we will have 32 bytes of extra code. We only need 82 deopts to lose the code size gains from making the stubs into builtins (ignoring the code size of the builtin).

With a quick look, the code that performs the call to the deoptimizer can be improved for arm64 (I believe the store to the isolate can be encoded in a single instruction, and we don't have to use x4 and x5 as scratch registers) but even then we'd have 6 instructions (so 16 bytes more than we currently have). The situation is similar for arm where each call to the deoptimizer needs 12 instructions with the change (instead of 3 currently).

Patch-set: 3