[ic] Remove the check for fast prototypes in LoadIC_Uninitialized
When handling load named properties (without feedback vectors) we used
to miss to runtimes if the prototypes aren't set. This was because we
wanted to give the prototype a chance to become fast, since most prototypes
start in slow mode but move to fast after the initial setup. Though this
check is not really useful when we don't have feedback vectors, and once
feedback vectors are allocated we will turn the prototypes fast anyway.

Bug:v8:8394, v8:8860
Change-Id: Ib2247e5e921f6375bda65310560ac832fd0339bf