Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> This shouldn't always miss though? The prototypes should turn fast pretty quickly? Is this avoided when running jitless somehow?
> 
I am sorry, my mistake I instrumented it incorrectly. Most of the cases that it misses into runtime are when there is no prototype (things like Arguments.length, Array.ForEach, Object.hasOwnProperty etc.,). The current code misses to runtime if it cannot load a valid prototype. So for Angular2-TypeScript, 98% of the loads that go through LoadIC_Uninitialized don't have a valid proptotype and they are missing into runtime.  There is also the Error function which has slow prototype for a very long time. May be because it is not used that much, it takes time to become fast? there aren't that many loads from Error though.

Would it be better if I still keep the check for fast prototype if the prototype exists but stop missing to runtime when there is no valid prototype? 

> lgtm anyway.

Patch-set: 2
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>