Update patch set 1
Patch Set 1:

Sorry for the delay, was at BlinkOn for the first half of this week.

> > Could you instead use a v8::Global<v8::Module> with a finalization callback (added with SetWeak()) to handle cleanup?
> 
> I'm not sure that is sufficient, but I may be wrong. My concern with this would just be that in the following:
> 
> const interval = setInterval(() => {
>   import('./asdf');
> });
> export function stop () {
>   // this function call is our lifetime point
>   clearInterval(interval);
> }
> 
> Consider that the module has no references from JS (so no registry reference). So no HostResolveImportedModule can resolve to the above module at all anymore. In theory then the v8::Module should be disposed right? Since it is entirely dereferenced, all we have at this point is the host_defined data being passed to each call of `import()`.
> 
> So our Weak callback would apply when the Module is disposed, which can be earlier than the point when its parent context is no longer unreachable through dynamic import (eg due to the interval being cleared).

I'm not sure I understand your example. On the web, the module imported as './asdf' would live as long as the context from which import() was invoked. Clearing the interval wouldn't get rid of it. Is that not true in the Node.js case?

Patch-set: 1
Reviewer: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>